requirements.txt
streamlit
langchain
openai
langchain-google-genai
langchain-anthropic
PyPDF2
python-docx